# Persistent L2 cache opt for cuda backend