@borg323 2022-12-16
We may have to revise the default for `cache_opt` but for now `true` is fine for wider testing.